refactor: models levados a pasta correta
