+++ app/Http/Controllers/Menu/FornecedorController.php
@@ -6,7 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
-use App\Models\Fornecedor;
+use App\Models\Menu\Fornecedor;
 
 class FornecedorController extends Controller
 {
