+++ app/Http/Controllers/Menu/ProdutoController.php
@@ -6,7 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
-use App\Models\Produto;
+use App\Models\Menu\Produto;
 
 class ProdutoController extends Controller
 {
