+++ app/Http/Controllers/Menu/SaidaProdutoController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers\Menu;
 
 use App\Http\Controllers\Controller;
-use App\Models\SaidaProduto;
+use App\Models\Menu\SaidaProduto;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
