+++ app/Models/Menu/SaidaProduto.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Menu;
 
 use Illuminate\Database\Eloquent\Model;
